Move core validators into a subpackage
