+++ datahub/core/validators/rules_based.py
@@ -5,137 +5,9 @@ from operator import contains, eq
 from typing import Any, Callable, Iterable, Sequence
 
 from rest_framework import serializers
-from rest_framework.exceptions import ValidationError
 from rest_framework.settings import api_settings
 
 from datahub.core.validate_utils import DataCombiner, is_blank, is_not_blank
-
-
-class RequiredUnlessAlreadyBlankValidator:
-    """
-    Class-level DRF validator for required fields that are allowed to stay null if already
-    null.
-
-    (Because of how validation works in DRF, this cannot be done as a field-level validator.)
-    """
-
-    required_message = 'This field is required.'
-
-    def __init__(self, *fields):
-        """
-        Initialises the validator with a list of fields to individually validate.
-
-        :param fields:  Fields that should be required (when not already null)
-        """
-        self.fields = fields
-        self.instance = None
-        self.partial = None
-
-    def __call__(self, attrs):
-        """Performs validation (called by DRF)."""
-        errors = {}
-        for field in self.fields:
-            if self.instance and is_blank(getattr(self.instance, field)):
-                continue
-
-            if self.partial and field not in attrs:
-                continue
-
-            if is_blank(attrs.get(field)):
-                errors[field] = self.required_message
-
-        if errors:
-            raise serializers.ValidationError(errors)
-
-    def set_context(self, serializer):
-        """
-        Saves a reference to the model instance and whether this is a partial update.
-
-        Called by DRF.
-        """
-        self.instance = serializer.instance
-        self.partial = serializer.partial
-
-    def __repr__(self):
-        """Returns the string representation of this object."""
-        return f'{self.__class__.__name__}(*{self.fields!r})'
-
-
-class AddressValidator:
-    """Validator for addresses."""
-
-    message = 'This field is required.'
-
-    DEFAULT_FIELDS_MAPPING = {
-        'address_1': {'required': True},
-        'address_2': {'required': False},
-        'address_town': {'required': True},
-        'address_county': {'required': False},
-        'address_postcode': {'required': False},
-        'address_country': {'required': True},
-    }
-
-    def __init__(self, lazy=False, fields_mapping=None):
-        """
-        Init the params.
-
-        :param lazy: True if you want to skip validation when none of the fields are set.
-            Useful when validating an extra and optional address where some fields
-            become required only if any of the fields are set.
-        :fields_mapping: dict with the field as a key and the value as a dict with
-            `required` == True or False
-        """
-        self.lazy = lazy
-        if fields_mapping:
-            self.fields_mapping = fields_mapping
-        else:
-            self.fields_mapping = self.DEFAULT_FIELDS_MAPPING
-        self.instance = None
-
-    def set_context(self, serializer):
-        """
-        This hook is called by the serializer instance,
-        prior to the validation call being made.
-        """
-        self.instance = getattr(serializer, 'instance', None)
-
-    def _should_validate(self, data_combined):
-        """
-        :returns: True if the data should be validated.
-            If lazy == True, the data should always be validated
-            If lazy == False, validate only if at least one field is set
-        """
-        if not self.lazy:
-            return True
-        return any(data_combined.values())
-
-    def _validate_fields(self, data_combined):
-        """
-        :returns: a dict containing potential errors
-        """
-        errors = {}
-        for field_name, mapping in self.fields_mapping.items():
-            if not mapping['required']:
-                continue
-            if not data_combined.get(field_name):
-                errors[field_name] = [self.message]
-        return errors
-
-    def __call__(self, data):
-        """Validate the address fields."""
-        data_combiner = DataCombiner(self.instance, data)
-
-        data_combined = {
-            field_name: data_combiner.get_value(field_name)
-            for field_name in self.fields_mapping.keys()
-        }
-
-        if not self._should_validate(data_combined):
-            return
-
-        errors = self._validate_fields(data_combined)
-        if errors:
-            raise ValidationError(errors)
 
 
 class AbstractRule(ABC):
